outputNote에 따라 악보그리기

+++ cviano/cviano/Key.h
@@ -82,7 +82,7 @@ namespace kb{
 void setWhiteKeyVector(cv::Mat& src, cv::Mat& dst, std::vector<kb::Key>& keys, cv::Rect rect);
 void sihyun(std::vector<kb::Key>, cv::VideoCapture,std::vector<std::pair<int,int>>&);
 void testforRoiMask(cv::Mat source, std::vector<kb::Key>& keys);
-void jinsoo(std::vector<std::pair<int, int>>& preNote);
+void jinsoo(std::vector<std::pair<int, float>>& preNote);
 
 
 #endif // !KEYBOARD
